requirements.txt
ansi2html==1.8.0
bcrypt==4.0.1
black==23.7.0
blessed==1.20.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
docker==6.1.3
fire==0.5.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
grpcio==1.60.0
idna==3.4
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
paramiko==3.3.1
pathspec==0.11.2
platformdirs==3.10.0
protobuf==4.25.1
pycparser==2.21
pyflakes==3.1.0
PyNaCl==1.5.0
python-editor==1.0.4
pywgkey==1.0.0
randomname==0.2.1
readchar==4.0.5
requests==2.31.0
-e git+ssh://git@github.com/Tkd-Alex/cosmospy-protobuf.git@1e219833084f7f8607e60f844ece59c56e7d7764#egg=sentinel_protobuf
six==1.16.0
SQLAlchemy==2.0.20
termcolor==2.3.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
websocket-client==1.6.2
Werkzeug==2.3.7
